feature: add theme provider to index.js, add theme context to navbar nad siderbar
